LayoutTests/imported/w3c/web-platform-tests/storage/persisted.https.any-expected.txt


PASS persisted() method exists and returns a Promise
FAIL persisted() returns a promise and resolves as boolean with false assert_equals: expected false but got true
